backend/requirements.txt
fastapi>=0.110.0,<1.0.0
uvicorn>=0.29.0,<1.0.0
python-dotenv>=1.0.0,<2.0.0
youtube-transcript-api>=1.0.0,<2.0.0
requests>=2.31.0,<3.0.0
openai>=1.0.0,<2.0.0
pydantic>=2.0.0,<3.0.0
